backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
gunicorn
psycopg2-binary
python-dotenv
dj-database-url
django-cors-headers
